webserver/requirements.txt
fastapi==0.103.2
pydantic==2.4.2
uvicorn==0.23.2
cassandra-driver==3.28.0
python-dotenv==1.0.0
pydantic-settings==2.0.3